+++ Liersch.Json.Demo/ExamplesWithReflection.cs
@@ -35,7 +35,7 @@ namespace Liersch.Json
       e1.StringValue="Example Text";
       e1.NotSerializedString="Other Text";
 
-      string json=new SLJsonSerializer().SerializeObject(e1).ToString();
+      string json=new SLJsonSerializer().Serialize(e1);
       Example e2=new SLJsonDeserializer().Deserialize<Example>(json);
 
       string f="{0,-24} => {1,16} - {2}";